requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
catboost

fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
pymysql